Added flush to beginning of iptables rules
This allows all rules to be flushed prior to restoring. This will ensure
that there are no lingering rules which should not be present. This
could possibly cause a brief disruption but should be fine. So this
functionality may need to be addressed at a later time if issues arise.